java/src/main/java/com/techelevator/dao/JdbcParkDao.java
package com.techelevator.dao;

import com.techelevator.exception.DaoException;
import com.techelevator.model.Park;
import org.springframework.dao.DataIntegrityViolationException;
import org.springframework.jdbc.CannotGetJdbcConnectionException;
import org.springframework.jdbc.core.JdbcTemplate;
import org.springframework.jdbc.support.rowset.SqlRowSet;

import javax.sql.DataSource;
import java.util.ArrayList;
import java.util.List;

public class JdbcParkDao implements ParkDao {

    private JdbcTemplate jdbcTemplate;

    public JdbcParkDao(DataSource dataSource) {
        this.jdbcTemplate = new JdbcTemplate(dataSource);
    }

    @Override
    public List<Park> getParks() {
    List<Park> parks = new ArrayList<>();
    String sql = "SELECT * FROM park;";
    try{
        SqlRowSet result = jdbcTemplate.queryForRowSet(sql);
        while(result.next()){
            Park park = mapRowToPark(result);
            parks.add(park);
        }
    } catch(CannotGetJdbcConnectionException e){
        throw new DaoException("Unable to connect to server or database");
    } catch(DataIntegrityViolationException e){
        throw new DaoException(("Data integrity violation"));
    }
    return parks;
    }

    private Park mapRowToPark(SqlRowSet results) {
        Park park = new Park();
        park.setParkId(results.getInt("park_id"));
        park.setName(results.getString("name"));
        park.setLocation(results.getString("location"));
        park.setEstablishDate(results.getDate("establish_date").toLocalDate());
        park.setArea(results.getInt("area"));
        park.setVisitors(results.getInt("visitors"));
        park.setDescription(results.getString("description"));
        return park;
    }

}
